requirements.txt
Jinja2==2.11.2
django-compressor==2.4
django-libsass==0.8
django-sass-processor==0.8.2